Add files via upload
Find the email address from the line starting with "From" (not "From:") from the "mbox-short.txt" file using a regex. Then count the number of unique email addresses and store them in a dictionary with the key being the unique email address and the value being the number of the unique email addresses. Then sorting and displaying the email addresses from the largest to the smallest.