Add files via upload
Open the file "mbox-short.txt" and save it in the list of all sender emails in the line starting with the word "From" (not "From:"). count the number of lines (messages) of each email in dictionary format with key: email address and val: number of messages. With the help of the list of tuples, print the big 3 email addresses who sent the most messages along with the number of messages.