Update 12220103_hitung_jumlah_email.py
Open the file "mbox-short.txt" and save it to a list of all sending emails in the line that starts with the word "From " (not "From:"). Count the number of lines (messages) of each email in dictionary format with key: email address and val: number of messages. With the help of the list of tuples, print the big 3 email addresses who sent the most messages along with the number of messages.