Create romeo-short_dictionary.py
Count the number of unique words in the romeo-short.txt file and save it in dictionary form with key in the form of unique words and value in the form of the number of unique words. Same word but with letters capitalization/lowercase and punctuation are considered the same (all lowercase and without punctuation). Show dictionary and count the number of elements in it (can use the len() command).